Revert "ProvidedDependencyParser"
This reverts commit 622e67a16ba28f170015425a953e4550485f2f99.
